DAQLaunch.py randomly doesn't return
Re-worked ParallelShell.py to not use threads and rather use
subprocess.py which I've plucked from python 2.4 (it promised to be
backwards compatible to python 2.2).  ParallelShell.py has *slightly*
different semantics but for the most part are identical to it's
previous behavior. DAQLaunch.py required a few small changes for this
and DeployPDAQ.py got a bit more, but have essentially the same interface.